refactor(ir): isolate callable component staging
Move cross-source prepared callable component lowering out of the codegen driver into its bounded backend adapter. Keep checker and identity decisions in the existing planning layer while passing frozen graph evidence and exact source plans into the staging boundary.

This preserves the M1A eligibility lane and makes the component transaction independently testable before the remaining ABI alias and atomic-publication work.

# Codex session
